+++ src/main/java/spharos/msg/domain/cart/service/CartProductService.java
@@ -3,7 +3,6 @@ package spharos.msg.domain.cart.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import spharos.msg.domain.cart.dto.CartProductCheckDto;
 import spharos.msg.domain.cart.dto.CartProductOptionResponseDto;
 import spharos.msg.domain.cart.dto.CartProductQuantityDto;
 import spharos.msg.domain.cart.dto.CartProductResponseDto;
@@ -16,7 +15,6 @@ import spharos.msg.domain.product.repository.ProductRepository;
 import spharos.msg.domain.users.entity.Users;
 import spharos.msg.domain.users.repository.UsersRepository;
 import spharos.msg.global.api.ApiResponse;
-import spharos.msg.global.api.code.status.ErrorStatus;
 import spharos.msg.global.api.code.status.SuccessStatus;
 
 import java.util.List;
@@ -75,41 +73,6 @@ public class CartProductService {
                         .collect(Collectors.toList()));
     }
 
-    @Transactional
-    public ApiResponse<?> updateCartProductOption(Long productOptionId, Long cartId, String userUuid) {
-        CartProduct cartProduct = cartProductRepository.findById(cartId).orElseThrow();
-        ProductOption productOption = productOptionRepository.findById(productOptionId).orElseThrow();
-        if(userCheck(cartProduct,userUuid)) {
-            cartProduct.updateCartProductOption(productOption);
-            return ApiResponse.of(SuccessStatus.CART_PRODUCT_UPDATE_SUCCESS,
-                    new CartProductResponseDto(cartProduct));
-        }
-        return ApiResponse.onFailure(ErrorStatus.NOT_CART_OWNER,null);
-    }
-
-    @Transactional
-    public ApiResponse<?> addCartProductQuantity(Long cartId, String userUuid) {
-        CartProduct cartProduct = cartProductRepository.findById(cartId).orElseThrow();
-        if(userCheck(cartProduct,userUuid)){
-            cartProduct.addOneCartProductQuantity();
-            return ApiResponse.of(SuccessStatus.CART_PRODUCT_UPDATE_SUCCESS,
-                    new CartProductResponseDto(cartProduct));
-        }
-        return ApiResponse.onFailure(ErrorStatus.NOT_CART_OWNER,null);
-    }
-
-    public ApiResponse<?> minusCartProductQuantity(Long cartId, String userUuid) {
-        return null;
-    }
-
-    public ApiResponse<?> checkCartProduct(CartProductCheckDto cartProductCheckDto, Long cartId, String userUuid) {
-        return null;
-    }
-
-    public ApiResponse<?> notCheckCartProduct(CartProductCheckDto cartProductCheckDto, Long cartId, String userUuid) {
-    return null;
-    }
-
     private ApiResponse<?> addCart(Users users, Long productOptionId, ProductOption productOption, Integer productQuantity) {
         List<CartProduct> cartProducts = cartProductRepository.findByUsers(users);
         for (CartProduct cartProduct : cartProducts) {
@@ -126,8 +89,5 @@ public class CartProductService {
                 .build();
         cartProductRepository.save(cartProduct);
         return ApiResponse.of(SuccessStatus.CART_PRODUCT_ADD_SUCCESS, null);
-    }
-    private boolean userCheck(CartProduct cartProduct, String userUuid){
-        return cartProduct.getUsers().getUuid().equals(userUuid);
     }
 }
